requirements.txt
discord.py==2.5.2
google-generativeai==0.8.5
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.26.0